third_party/blink/web_tests/animations/svg-attribute-interpolation/svg-startOffset-interpolation-002-expected.txt

This is a testharness.js-based test.
[FAIL] SMIL: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0) is [250]
  assert_equals: expected "250 " but got "35 "
[FAIL] SMIL: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0.2) is [450]
  assert_equals: expected "450 " but got "278 "
[FAIL] SMIL: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0.6) is [850]
  assert_equals: expected "850 " but got "764 "
[FAIL] Web Animations: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (-0.4) is [-150]
  assert_equals: expected "- 150 " but got "- 451 "
[FAIL] Web Animations: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0) is [250]
  assert_equals: expected "250 " but got "35 "
[FAIL] Web Animations: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0.2) is [450]
  assert_equals: expected "450 " but got "278 "
[FAIL] Web Animations: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (0.6) is [850]
  assert_equals: expected "850 " but got "764 "
[FAIL] Web Animations: Interpolate attribute <startOffset> from replace [25%] to replace [25em] at (1.4) is [1650]
  assert_equals: expected "1650 " but got "1736 "
Harness: the test ran to completion.
